Make dit_participants writable in interaction API
This will replace the `dit_adviser` and `dit_team` fields. `dit_adviser` and `dit_team` are retained for temporary backwards compatibility.

The `dit_participants` model field is used as the data source for all three serialiser fields. However, the `dit_adviser` and `dit_team` are still kept up to date to not break other services that are consuming our database.

The backwards compatibility adds some complexity to the logic, but this will be simplified as soon as that backwards compatibility is removed.